move the component to other files
